wscript: remove 32-bit host environment
It's 2025 now -- there is no need to distribute binaries for 32-bit hosts
anymore, and it's come to the point where some machines (like my laptop!)
cannot seem to successfully compile a `-m32` binary anymore.  And for that
matter, modern macOS cannot even *run* 32-bit Intel binaries.

Extirpate the '32bit' build environment, and replace all uses with 'local'
(and, while we are at it, fix the build for pdc2png).

Signed-off-by: Joshua Wise <joshua@joshuawise.com>